Factor out list support, and implement opening/closing
This PR factors out the loclist logic to be more generic and support the
quickfix list (as well as moving it to its own file). It also implements
the opening and closing of said lists, with logic to ensure we don't
close the list when another plugin is using it.

I am not sure what the best way to address preferences is, as I wanted
to maintain backward compaibility. I think the implementation here is
acceptable, but @w0rp's opinion may differ.

The settings still need to be documented, but I will follow up on that
after the settings interface is finalized.

Closes #136, #45