Take care of the `/` case when joining paths
The current code will create a path like `//bin/activate` when joining
the parent directory `/` and `/bin/activate`. This triggers a chain
reaction that makes Vim unresponsive for several seconds when run in an
MSYS2 environment on Windows due to special handling of the `//` path,
and might potentially (but very exotically) lead to unexpected results
even on POSIX systems since the interpretation of initial `//` paths is
implementation dependent.

Note that this only takes care of a single case: at a quick glance there
are similar issues in `ale#python#FindProjectRoot` in
`autoload/ale/python.vim` as well as in `autoload/ale/c.vim`. I am using
this "quick fix" as a starting point for a discussion on how it should
be handled in general.

[0]: http://pubs.opengroup.org/onlinepubs/009695399/basedefs/xbd_chap04.html#tag_04_11