Refactor
Extraced function to create extra error message
Extracted function that parses the flow error for an extra error
Removed the overly cautious if statement at the beginning of the regular error parsing. It is now back to the way it was before I added it. It's prob safe to assume that if flow has errors, each error has a message prop unless flow is seriously broken.