fix issue with go1.10 govet
In go1.10 they have extended 'go vet' with access to more information.
As a side effect it seems you can no longer run 'go vet' against a
single file. You now need to specify all the files or the package.
Using 'go tool vet' you can run it against a single file. Using it is
discouraged, but it is supported.

Similar to w0rp/ale#936, without this it fails to find anything
(types, functions, etc) the other files in the package and reports a
bunch of spurious errors.

This change adds a variable to run 'go vet' over the package containing
the file loaded in the current buffer or, if not set, it will run 'go
tool vet' over the current buffer.