Show multiple lints in echo status and ALEDetail
I got annoyed that I'd need to dig through :llist to see the multiple
errors on a given line. I wanted ALE to show me all the problems, and
show the most critical errors first.

So that's what it now does!

By reimplementing FindItemAtCursor and refactoring the BinarySearch, we
now show all lints for a given line, or if the user has their cursor
exactly on the line/col of a highlighted lint we show all the lints for
that given line/col. We also sort the output based on decreasing error type
severity.

I left the old BinarySearch because that's what shows the popup
balloons, and I figure the old functionality suffices for the balloons.

Resolves #1426