Add better support for Haskell stack compiler tools
`ale#Escape` does not play nicely with a common pattern in haskell
executable usage. Many `stack` users invoke compiler tools via
`stack exec [command] -- [options]`. `ale#Escape` wraps this command in
quotes, because it includes spaces. This results in silent failure of
the fixer.

This commit adds support for `stack` as the executable of a tool. This
follows a pattern that has been implemented for `bundler`'s tool chain.