Add fixer for Go modules
Go 1.11 introduces the "go.mod" file and "go mod edit -fmt" to format
it. This adds a fixer for that.

This requires Go 1.11 to test, and a recently new version of vim-go
which sets the `filetype` for this file (was merged today). Or you can
stick this in your vimrc:

	autocmd BufRead,BufNewFile go.mod set ft=gomod

I'll see if we can get it in the standard runtime files too.

Also allow configuring the "go" executable with `ale_go_go_executable`;
this has come up occasionally.