Call QuickFixCmdPre and QuickFixCmdPost autocmds
The concrete problem that this solves is that I want to do "stuff" when
there are quickfix/location list errors. I initially tried hooking in to
ALELintPost, but that doesn't work as ALE sets errors asynchronously.

By hooking in to QuickFixCmd* this can be done in a generic (non-ALE
specific) way, that will work for e.g. :make as well; for example to
switch the colour of the statusline depending on the quickfix errors:

    fun! s:errcolor()
        if len(getloclist(bufnr('%'))) > 0 || len(getqflist()) > 0
            hi StatusLine term=bold,reverse cterm=bold,reverse gui=bold,reverse
                        \ ctermfg=1 guifg=#cd0000
        else
            hi StatusLine term=bold,reverse cterm=bold,reverse gui=bold,reverse
                        \ ctermfg=NONE guifg=NONE
        endif
    endfun

    autocmd QuickFixCmdPost ?expr,*make call s:errcolor() | redrawstatus

IMHO this is a more elegant/Vim-native solution than using ALELintPost
anyway (I saw there is an undocumented setting for excecuting it in
sync).

I'll appropriate documentation and tests if the approach in this PR is
approved (no need to needlessly write docs/struggle with vader).