Default LSP linter root to the top of the Git repository
Currently, we detect the linter root based on a variety of techniques.
However, these techniques are not foolproof. For example, clangd works
fine for many things without a compile_commands.json file, and Go
projects may be built outside of the GOPATH to take advantage of Go
1.11's automatic module support.

If the language-specific technique does not produce a root, then the LSP
linter is not enabled, which is suboptimal. As a fallback, detect
whether we are in a Git repository, and if so, use that as the project
root. Handle both regular repositories and submodules.