Ruby/Brakeman: only lint current file
When the Brakeman linter was added, Brakeman didn't yet have the ability
to only check a single file. This results in ALE filling the location
list with violations from elsewhere in the project that aren't relevant
to the current buffer.

This change uses the `--only-files` argument to scope the Brakeman check
to the file of the current buffer. Note that this change will cause some
controller/view interactions to be missed, as Brakeman won't consider
violations that require scanning the current buffer's file in the
context of other files in the project (such as XSS with a model fetched
in a controller but output unescaped in a view), but it cuts down
significantly on unactionable noise.