Support additional language servers in Go
go-langserver[1] has long been the "go to" language server for Go, but
recently work on it has (in their words) been "deprioritized" in favour
of the upcoming official Go LSP (golsp[2]). Unfortunately, this isn't
really ready yet, so for now the bingo[3] fork recommended

Instead of adding two new LSP linters, this modifies the existing one so
that setting g:ale_go_langserver_executable to "bingo" or "golsp" will
work as expected; this didn't work previously due to muckery with adding
a flag.

It also fixes the ale#go#FindProjectRoot() function, as this would get
the GOPATH location, not the project location.

At some point in the future, the default should be changed, once
everything has stabilized.

1 https://github.com/sourcegraph/go-langserver
2 https://github.com/golang/go/wiki/GoLSP:-Go-Language-Smartness-Provider
3 https://github.com/saibing/bingo