Call rstcheck with --config option
The --config option is set to the original location of the rst file that
is being linted. This means that `rstcheck` will search for a config
file walking up from that location, instead of the location of the
temporary file that ALE generates.

Closes #2408. The resolution is based on recent changes in `rstcheck`,
cf. https://github.com/myint/rstcheck/issues/56.
There is a new option `g:ale_rst_rstcheck_use_project_config` that
enables this feature. This is so that anyone using an older
version <= 3.3.1 of `rstcheck`, which does not have the `--config`
options, is still able to run the linter. Since
https://github.com/myint/rstcheck/pull/57 has been merged,
any newer version of `rstcheck` should work out of the box.

The value `g:ale_rst_rstcheck_use_project_config` is 1 ("on") by
default, because I think people would generally expect their project
configuration to be taken into account in the default configuration.

A new option `g:ale_rst_rstcheck_options` is also provided. This
closes #2144. However, the main intended useage of this option is to
pass `--debug` to `rstcheck` (showing which config file is being used).

The documentation has been updated with a new
`ale-restructuredtext-rstcheck` section.