Tweak cppcheck linter handling.
If a file is modified, using compile_commands.json doesn't work, because
it doesn't take into account the local changes. Instead, include the
file's directory in the cppcheck call with -I. Also do this if
compile_commands.json isn't found.

This could be expanded to add other -I flags to cppcheck based on the
command in compile_commands.json, if found.