Improve java checkstyle command configuration.
The g:ale_java_checkstyle_options only worked with absolute paths making
it difficult to have a per project configuration. This MR introduces a
new g:ale_java_checkstyle_config configuration. This option can be an
absolute path or a filename. When a filename is set then the nearest
file that matches it in the current buffer folder or any of its parents
will be used.

Note: The above is a breaking change since it replaces the
g:ale_java_checkstyle_options variable with a different one.

Also a new configuration g:ale_java_checkstyle_executable is added. This
allows users to create more specialized launch scripts (#1305) to invoke
checkstyle.