clang: pass appropriate language for header files
Some C/C++ constructs cause errors in translation units, but not in
header files. For example, the following:

    static const int x = 0;

...fails to compile with `clang -x c -Wall -Werror -fsyntax-only`, but
succeeds with `clang -x c-header -Wall -Werror -fsyntax-only`.

This commit updates the Clang linters to use a buffer's filetype and
extension to pass the appropriate language as `-x`. This fixes
false-positive errors when linting header files.