Map mypy "note" messages to info-level lints
ALE shows errors from mypy, but not notes. This firstly makes it hard to
debug some issues, since mypy often emits a note immediately after an
error with more explanation of the types involved or a link to an
explanation of the error (and these notes aren't even visible in
:ALEDetail), and secondly means that adding reveal_type to something
doesn't show the type of the expression in the editor – you have to go
and run mypy by hand to see the result.

We map "note" messages to ALEInfo highlight level groups because that
means when the user runs :lwin they'll see the expanded note if there's
also an error on the line, and if there's no error they'll just see the
ALEInfo highlight group.

Fixes #2704